+++ comfortbox-api-server/comfortbox-api-server.js
@@ -1,0 +1,12 @@
+module.exports = function(RED) {
+  'use strict';
+
+  function ComfortboxApiServerNode(config) {
+    RED.nodes.createNode(this, config);
+    this.host = config.host;
+    this.port = config.port;
+    this.protocol = config.protocol;
+    this.accessToken = config.accessToken;
+  }
+  RED.nodes.registerType("comfortbox-api-server", ComfortboxApiServerNode);
+}