+++ CHANGELOG.md
@@ -7,5 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - Initial setup for packaging a npm module
+- Config `comfortbox-api-server`
+- Node `display text`
 
 [Unreleased]: https://github.com/dwettstein/node-red-contrib-comfortbox/compare/master...HEAD